codex(refactor): extract compose section helpers
