+++ newsletter_core/application/generation/compose.py
@@ -8,18 +8,25 @@ from typing import Any, Dict, List
 import yaml
 from jinja2 import Environment, FileSystemLoader, select_autoescape
 
-from newsletter.date_utils import (
-    extract_source_and_date,
-    format_date_for_display,
-    standardize_date,
-)
+from newsletter.date_utils import extract_source_and_date, format_date_for_display
 from newsletter.utils.logger import get_logger
 
 from .compose_inputs import (
     NewsletterConfig,
-    extract_test_config,
     normalize_compose_input,
     resolve_style_config,
+)
+from .compose_sections import (
+    add_emoji_to_section_title,
+    create_grouped_sections,
+    extract_and_prepare_top_articles,
+    extract_definitions,
+    extract_food_for_thought,
+    extract_key_definitions_for_compact,
+    extract_top_articles_from_sections,
+    format_compact_source_date,
+    prepare_grouped_sections_for_compact,
+    prepare_top_articles_for_compact,
 )
 
 # 로거 초기화
@@ -114,192 +121,6 @@ def compose_newsletter(data: Any, template_dir: str, style: str = "detailed") ->
         definitions,
         food_for_thought,
     )
-
-
-def extract_and_prepare_top_articles(
-    data: Dict[str, Any], count: int = 3
-) -> List[Dict[str, Any]]:
-    """
-    상위 기사(top_articles)를 sections/grouped_sections와 분리하여 추출 및 템플릿용으로 준비합니다.
-    - data["top_articles"]가 있으면 우선 사용
-    - 없으면 data["sections"]에서 추출
-    - 항상 top_articles는 grouped_sections와 별도 관리
-    """
-    # 기존 top_articles가 있으면 사용
-    if "top_articles" in data and data["top_articles"]:
-        top_articles = data["top_articles"][:count]
-    elif "sections" in data:
-        # sections에서 첫 번째 기사들 추출
-        top_articles = extract_top_articles_from_sections(data["sections"])[:count]
-    else:
-        # sections도 없고 top_articles도 없는 경우 빈 리스트 반환
-        top_articles = []
-
-    # 템플릿용 포맷팅
-    prepared_articles = []
-    for article in top_articles:
-        # 날짜 형식 포맷팅
-        source_and_date = article.get("source_and_date", "")
-        if source_and_date:
-            source, date_str = extract_source_and_date(source_and_date)
-            if date_str:
-                formatted_date = format_date_for_display(date_str=date_str)
-                if formatted_date:
-                    source_and_date = f"{source} · {formatted_date}"
-
-        prepared_article = {
-            "title": article.get("title", ""),
-            "url": article.get("url", "#"),
-            "snippet": article.get("snippet", article.get("summary_text", "")),
-            "source_and_date": source_and_date,
-        }
-        prepared_articles.append(prepared_article)
-
-    return prepared_articles
-
-
-def create_grouped_sections(
-    data: Dict[str, Any],
-    top_articles: List[Dict[str, Any]],
-    max_groups: int = 6,
-    max_articles: int = None,
-) -> List[Dict[str, Any]]:
-    """
-    grouped_sections(주제별 기사 그룹)은 top_articles와 분리하여 생성합니다.
-    - top_articles에 포함된 기사 URL은 제외
-    - 항상 top_articles와 별도 관리
-    - 각 그룹에 해당하는 definitions 포함
-    """
-    # 기존 grouped_sections가 있으면 사용
-    if "grouped_sections" in data:
-        existing_sections = data["grouped_sections"][:max_groups]
-        # 기존 grouped_sections에 definitions가 없으면 추가
-        for section in existing_sections:
-            if "definitions" not in section or not section["definitions"]:
-                # 원본 sections에서 해당하는 정의 찾기
-                section_title = (
-                    section.get("heading", "")
-                    .replace("🚀 ", "")
-                    .replace("🏛️ ", "")
-                    .replace("📊 ", "")
-                    .replace("🌐 ", "")
-                    .replace("📈 ", "")
-                )
-                original_section = next(
-                    (
-                        s
-                        for s in data.get("sections", [])
-                        if s.get("title") == section_title
-                    ),
-                    None,
-                )
-                if original_section and "definitions" in original_section:
-                    section["definitions"] = original_section["definitions"]
-                else:
-                    section["definitions"] = []
-        return existing_sections
-
-    # sections에서 grouped_sections 생성
-    sections = data.get("sections", [])
-    if not sections:
-        return []
-
-    # 이미 선택된 상위 기사들의 URL 추출
-    excluded_urls = {article.get("url", "") for article in top_articles}
-
-    grouped_sections = []
-    article_count = 0
-
-    # compact 모드 여부 확인 (max_articles가 설정되어 있으면 compact 모드)
-    is_compact = max_articles is not None
-
-    for section in sections[:max_groups]:
-        # 섹션의 기사 목록 가져오기
-        news_links = section.get("news_links", [])
-        articles = section.get("articles", [])
-        article_list = news_links if news_links else articles
-
-        # 상위 기사로 선택된 것들 제외
-        remaining_articles = [
-            link for link in article_list if link.get("url", "") not in excluded_urls
-        ]
-
-        # max_articles 제한 확인 (compact 모드용)
-        if max_articles and article_count + len(
-            remaining_articles
-        ) > max_articles - len(top_articles):
-            remaining_articles = remaining_articles[
-                : max_articles - len(top_articles) - article_count
-            ]
-
-        if remaining_articles:  # 남은 기사가 있을 때만 섹션 추가
-            # 이모지 추가된 섹션 제목
-            section_title = add_emoji_to_section_title(section.get("title", "기타"))
-
-            # intro 생성 - compact 모드에서는 간결하게
-            intro = ""
-            if is_compact:
-                # compact 모드: section.intro를 직접 사용 (요약 체인에서 이미 생성됨)
-                intro = section.get("intro", "")
-                # intro가 없으면 기본 메시지 생성
-                if not intro:
-                    intro = f"{section.get('title', '기타')}에 대한 주요 동향입니다."
-            else:
-                # detailed 모드: 기존 방식 유지
-                intro = (
-                    section.get("summary_paragraphs", [""])[0]
-                    if section.get("summary_paragraphs")
-                    else ""
-                )
-
-            grouped_section = {
-                "heading": section_title,
-                "intro": intro,
-                "articles": remaining_articles,
-                "definitions": section.get("definitions", []),  # 섹션의 정의 포함
-            }
-            grouped_sections.append(grouped_section)
-            article_count += len(remaining_articles)
-
-            # max_articles 도달하면 중단
-            if max_articles and article_count >= max_articles - len(top_articles):
-                break
-
-    return grouped_sections
-
-
-def extract_definitions(
-    data: Dict[str, Any], grouped_sections: List[Dict[str, Any]], config: Dict[str, Any]
-) -> List[Dict[str, str]]:
-    """용어 정의들을 추출"""
-
-    # 기존 definitions가 있으면 사용
-    if "definitions" in data:
-        definitions = data["definitions"]
-        if config["max_definitions"]:
-            return definitions[: config["max_definitions"]]
-        return definitions
-
-    # sections에서 정의 추출
-    sections = data.get("sections", [])
-    return extract_key_definitions_for_compact(sections)[
-        : config["max_definitions"] or 999
-    ]
-
-
-def extract_food_for_thought(data: Dict[str, Any]) -> Any:
-    """생각해볼 거리 추출"""
-    food_for_thought = data.get("food_for_thought")
-
-    if not food_for_thought:
-        return {"message": ""}
-
-    if isinstance(food_for_thought, dict):
-        # 딕셔너리 형태 그대로 반환
-        return food_for_thought
-    else:
-        # 문자열인 경우 딕셔너리로 변환하여 반환
-        return {"message": str(food_for_thought)}
 
 
 def render_newsletter_template(
@@ -644,194 +465,6 @@ def process_compact_newsletter_data(newsletter_data: Dict[str, Any]) -> Dict[str
             compact_data["food_for_thought"] = {"message": str(food_for_thought)}
 
     return compact_data
-
-
-def extract_top_articles_from_sections(
-    sections: List[Dict[str, Any]],
-) -> List[Dict[str, Any]]:
-    """
-    각 섹션에서 첫 번째 뉴스 링크를 추출하여 상위 기사로 만듭니다.
-    중복 URL 제거 로직 포함.
-    """
-    top_articles = []
-    seen_urls = set()  # 중복 URL 추적
-
-    for section in sections[:3]:  # 최대 3개 섹션에서
-        news_links = section.get("news_links", [])
-        if news_links:
-            # 섹션의 모든 기사를 확인하여 중복되지 않은 첫 번째 기사 선택
-            for article in news_links:
-                article_url = article.get("url", "#")
-
-                # URL이 이미 선택되었거나 유효하지 않으면 건너뛰기
-                if article_url in seen_urls or article_url == "#" or not article_url:
-                    continue
-
-                # 요약 텍스트 생성 (섹션의 첫 번째 요약 문단 사용)
-                summary_paragraphs = section.get("summary_paragraphs", [])
-                snippet = summary_paragraphs[0] if summary_paragraphs else ""
-
-                top_article = {
-                    "title": article.get("title", ""),
-                    "url": article_url,
-                    "snippet": snippet[:150] + "..." if len(snippet) > 150 else snippet,
-                    "source_and_date": article.get("source_and_date", ""),
-                }
-
-                top_articles.append(top_article)
-                seen_urls.add(article_url)
-                break  # 해당 섹션에서 하나만 선택하고 다음 섹션으로
-
-    return top_articles
-
-
-def prepare_top_articles_for_compact(
-    top_articles: List[Dict[str, Any]],
-) -> List[Dict[str, Any]]:
-    """
-    상위 기사들을 간결한 템플릿용으로 포맷팅합니다.
-    """
-    prepared_articles = []
-
-    for article in top_articles:
-        # 날짜 형식 포맷팅
-        source_and_date = article.get("source_and_date", "")
-        if source_and_date:
-            source, date_str = extract_source_and_date(source_and_date)
-            if date_str:
-                formatted_date = format_date_for_display(date_str=date_str)
-                if formatted_date:
-                    source_and_date = f"{source} · {formatted_date}"
-
-        prepared_article = {
-            "title": article.get("title", ""),
-            "url": article.get("url", "#"),
-            "snippet": article.get("snippet", article.get("summary_text", "")),
-            "source_and_date": source_and_date,
-        }
-        prepared_articles.append(prepared_article)
-
-    return prepared_articles
-
-
-def prepare_grouped_sections_for_compact(
-    sections: List[Dict[str, Any]], exclude_articles: List[Dict[str, Any]]
-) -> List[Dict[str, Any]]:
-    """
-    섹션들을 간결한 템플릿용 그룹으로 변환합니다.
-    """
-    grouped_sections = []
-
-    # 이미 상위 기사로 선택된 기사들의 URL 추출
-    excluded_urls = {article.get("url", "") for article in exclude_articles}
-
-    for section in sections:
-        # compact 모드에서는 news_links 대신 articles를 사용
-        news_links = section.get("news_links", [])
-        articles = section.get("articles", [])
-
-        # news_links가 있으면 사용하고, 없으면 articles 사용
-        article_list = news_links if news_links else articles
-
-        remaining_articles = [
-            link for link in article_list if link.get("url", "") not in excluded_urls
-        ]
-
-        if remaining_articles:  # 남은 기사가 있을 때만 섹션 추가
-            # 그룹 제목에 이모지 추가
-            section_title = section.get("title", "")
-            group_heading = add_emoji_to_section_title(section_title)
-
-            # 간단한 소개 문구 생성 (첫 번째 요약 문단의 첫 문장 사용)
-            # compact 모드에서는 intro 필드 사용
-            intro = section.get("intro", "")
-            if not intro:
-                summary_paragraphs = section.get("summary_paragraphs", [])
-                if summary_paragraphs:
-                    first_paragraph = summary_paragraphs[0]
-                    sentences = first_paragraph.split(". ")
-                    intro = (
-                        sentences[0] + "."
-                        if sentences
-                        else first_paragraph[:100] + "..."
-                    )
-
-            grouped_section = {
-                "heading": group_heading,
-                "intro": intro,
-                "articles": [
-                    {
-                        "title": article.get("title", ""),
-                        "url": article.get("url", "#"),
-                        "source_and_date": format_compact_source_date(
-                            article.get("source_and_date", "")
-                        ),
-                    }
-                    for article in remaining_articles[:4]  # 최대 4개까지만
-                ],
-            }
-            grouped_sections.append(grouped_section)
-
-    return grouped_sections
-
-
-def add_emoji_to_section_title(title: str) -> str:
-    """
-    섹션 제목에 적절한 이모지를 추가합니다.
-    """
-    title_lower = title.lower()
-
-    if any(word in title_lower for word in ["투자", "펀딩", "자금", "ipo", "상장"]):
-        return f"🚀 {title}"
-    elif any(word in title_lower for word in ["정책", "규제", "법", "윤리"]):
-        return f"🏛️ {title}"
-    elif any(word in title_lower for word in ["연구", "기술", "개발", "특허"]):
-        return f"📊 {title}"
-    elif any(word in title_lower for word in ["시장", "수요", "트렌드", "소비"]):
-        return f"🌐 {title}"
-    else:
-        return f"📈 {title}"
-
-
-def format_compact_source_date(source_and_date: str) -> str:
-    """
-    간결한 템플릿용으로 출처와 날짜를 포맷팅합니다.
-    """
-    if not source_and_date:
-        return ""
-
-    source, date_str = extract_source_and_date(source_and_date)
-    if date_str:
-        formatted_date = format_date_for_display(date_str=date_str)
-        if formatted_date:
-            return f"{source} · {formatted_date}"
-
-    return source_and_date
-
-
-def extract_key_definitions_for_compact(
-    sections: List[Dict[str, Any]],
-) -> List[Dict[str, str]]:
-    """
-    각 섹션에서 핵심 용어 정의를 추출합니다. 최대 3개까지만 반환합니다.
-
-    Args:
-        sections: 뉴스레터 섹션 리스트
-
-    Returns:
-        List[Dict]: 용어와 설명을 포함한 딕셔너리 리스트
-    """
-    all_definitions = []
-
-    for section in sections:
-        definitions = section.get("definitions", [])
-        if definitions:
-            # 각 섹션에서 첫 번째 정의만 가져옴 (가장 중요한 용어로 간주)
-            if len(definitions) > 0:
-                all_definitions.append(definitions[0])
-
-    # 최대 3개까지만 반환 (가독성을 위해)
-    return all_definitions[:3]
 
 
 def load_newsletter_settings(config_file: str = "config.yml") -> Dict[str, Any]:
